ci: separate workflow for fuzzing using corpus
Fetching the corpus in the main CI worflow is a problem because
it will fail when the azure storage is modified during the download:

> ERROR: The specified blob does not exist.
> RequestId:76c40723-901e-0098-38b9-8215c9000000
> Time:2023-05-09T21:03:10.8992962Z
> ErrorCode:BlobNotFound

This occurs when there is multiple runs at the same time. This
happens for example when two pull requests are opened.
This will cause CI to go red for no good reason.

With this patch, main CI will simply run the fuzz targets without
using the corpus, as a sanity check. Then, another workflow will
run at 03:12 AM UTC on Sunday, Monday, and Saturday using the corpus
from azure storage.

Future work: make the scheduled run fuzz longer and in parallel.