ci: coverage reporting in PRs
Adds a new workflow reporting code coverage when opening a pull request.

The idea is to help find blind spots in our test suites and inform where testing
efforts should go, but there is no hard requirement related to the numbers.
We are allowed to merge pull requests regardless of the coverage percentage,
and even if the coverage workflow fails.

Past coverage data is stored in a separate branch (cov-data).
It gets updated when the workflow run against master.