Connects the RDPDR static virtual channel
This simply connects the channel in a manner that apparently works
(based on the logs), nothing is yet sent over the channel.

Currently things are named "efs" based on [MS-RDPEFS], however arguably
they should be named "rdpdr" based on the fact that this virtual channel
is named "RDPDR" in the protocol, and other things (like smartcard, printers)
are redirected over it.