Adds utility structs for Vec<CapabilityMessage> and Vec<DeviceAnnounceHeader>
Unfortunately these structs rely on taking ownership of the underlying
data when it's used. This would work very nicely/efficiently, except
under normal circumstances ironrdp-client connects and then reconnects
during a session, which means that in the second run (the one that the user
actually interacts with) the data is gone, and empty lists are sent for
both of these values.

In the next commit I will fix this by changing take() to take_clone().