fix(connector): better security protocol selection
Replace the bitflag-based config API with a boolean-based one:

- `enable-winlogon`: set the PROTOCOL_SSL flag
- `enable-credssp`: set the PROTOCOL_HYBRID and PROTOCOL_HYBRID_EX flags

The `--security_protocol` argument was removed from the native client
CLI, and instead it’s possible to disable specific protocols with
`--no-winlogon` and `--no-credssp`. By default, both protocols are
enabled for maximum compatibility with most RDP servers. We may change
the defaults in the future.