Adds `DvcPduEncode` trait and `MonitorLayoutPdu`
`DvcPduEncode` trait simply inherits `PduEncode`. It is a marker trait
to indicate that the implementor, when encoded, is ready to be wrapped
in DataFirst and/or Data PDU(s).

`MonitorLayoutPdu` is a copy of the pdu by the same name in the
`ironrdp-pdu` crate. It is moved (well, copied for now) to the `ironrdp-dvc`
crate and implements `DvcPduEncode`.