Cleaning up the get_dvc_processor api
The get_dvc_processor api was a bit confusing, returning a
`Option<(&T, Option<DynamicChannelId>)>` with a list of semantics
for what None meant for the whole value and the `DynamicChannelId`.

This change converts what was previously `DynamicVirtualChannel` to
be named `DynamicVirtualChannelInternal`, and then adds a new public type
named `DynamicVirtualChannel` which wraps the internal type and provides
some more straightforward semantics via its function names.