fix(ffi): use patched diplomat
This patched version of Diplomat is backporting an important fix
from the more up to date versions of Diplomat.
Without this patch, the code generated by the diplomat::bridge macro
contains an easy to trigger UB.
This UB is triggered by the generated C# code when passing an empty
byte[] array.