fix(web): raise only `TERMINATED` message on session termination with the error backtrace
When the session ends with an error, `iron-remote-desktop` raises `ERROR` in pair with `TERMINATED`. Handling function in Devolutions Gateway Standalone handles these types of events similarly. In result, we see only the message from `TERMINATED` event in the `error notify bar`, because the message from `ERROR` is overwritten.

This commit fixes the issue by only raising the `TERMINATED` with the error backtrace.