fix(web): apply correct key codes to scan codes mapping
The root cause of the bug is irrelevant keycodes to scancodes mappings. These mappings were translating the browser's key codes (like "KeyA") to OS-specific scancode (OS where the browser runs). However, that's not exactly what we need to do.
We need to map the browser's key codes to Windows scancodes - the only format that is accepted by _VNC_ and _RDP) modules.
Let's look at an example for a better understanding. _Safari_ browser used _Linux_'s _Gecko) mappings as a fallback (because there were no MacOS-specific mappings). For a given key, iron-remote-desktop was providing a scancode that did not correspond to the Windows scancode for that same key. As a result, the `IronVNC` module incorrectly mapped this scancode to a `KeySym`, which resulted in an incorrect `KeySym` or `NO_SYMBOL`.

Fixes https://github.com/Devolutions/IronVNC/issues/882.