build(deps): replace opus by opus2
opus is unmaintained and ponits to a 4-year old commit of the opus
C library. This does not compile anymore on our CI, because their
CMakeList.txt requires an older version of cmake that is not available
in the runners we use. opus2 is a fork that points to a more recent
version of it.